[Submission]
When I first turned worked on this assignment, I did not know that there were more instructions on the "main" page, I only followed the instructions on the "read me" page so I assumed we just needed to implement the code on the catalan.py file and create a graph on the catalan.py file. I went back and finished the main, which uploads the data to the "data" file and the graphs to the figures file.